Joining of several commom functions in both examples
